Long -> Double

+++ src/main/java/com/datastax/kawoosh/analyser/rules/DroppedMutationsCheckRule.java
@@ -4,6 +4,6 @@ import com.datastax.kawoosh.analyser.ClusterConfigRetriver;
 
 public class DroppedMutationsCheckRule extends TableValueInRangeRule {
     public DroppedMutationsCheckRule(ClusterConfigRetriver clusterConfigRetriver) {
-        super(clusterConfigRetriver, "Dropped Mutations", "Dropped Mutations", (Long) 0L,(Long) 0L);
+        super(clusterConfigRetriver, "Dropped Mutations", "Dropped Mutations", 0d,0d);
     }
 }
