+++ src/main/java/com/datastax/kawoosh/analyser/rules/LargePartitionCheckRule.java
@@ -4,6 +4,6 @@ import com.datastax.kawoosh.analyser.ClusterConfigRetriver;
 
 public class LargePartitionCheckRule extends TableValueInRangeRule {
     public LargePartitionCheckRule(ClusterConfigRetriver clusterConfigRetriver) {
-        super(clusterConfigRetriver, "Large Partition", "Compacted partition maximum bytes", (Long) 0L,(Long) 100000000L);
+        super(clusterConfigRetriver, "Large Partition", "Compacted partition maximum bytes",  0d, 100000000d);
     }
 }
