+++ src/main/java/com/datastax/kawoosh/analyser/rules/TombstonesCheckRule.java
@@ -4,6 +4,6 @@ import com.datastax.kawoosh.analyser.ClusterConfigRetriver;
 
 public class TombstonesCheckRule extends TableValueInRangeRule {
     public TombstonesCheckRule(ClusterConfigRetriver clusterConfigRetriver) {
-        super(clusterConfigRetriver, "Tombstones", "Maximum tombstones per slice (last five minutes)", (Long) 0L,(Long) 1000L);
+        super(clusterConfigRetriver, "Tombstones", "Maximum tombstones per slice (last five minutes)",0d,1000d);
     }
 }
